+++ packages/assertions/.prettierignore
@@ -1,0 +1,1 @@
+dist/
